* Refactor "interfaces/controllers" directory structure.
