+++ src/interfaces/controllers/base/IBaseCharacterController.ts
