+++ src/interfaces/controllers/base/IBasePlayerCharacterController.ts
@@ -1,4 +1,4 @@
-import { IBaseCharacterControllerConstructorParams } from '@mmorpg/interfaces/controllers/character/IBaseCharacterController';
+import { IBaseCharacterControllerConstructorParams } from '@mmorpg/interfaces/controllers/base/IBaseCharacterController';
 import PlayerCharacter from '@mmorpg/game-objects/characters/PlayerCharacter';
 
 interface IBasePlayerCharacterControllerConstructorParams extends IBaseCharacterControllerConstructorParams {
